Move method get to Header
