--- core/src/main/resources/lib/layout/pageHeader.jelly.orig
@@ -1,111 +1,0 @@
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:i="jelly:fmt" xmlns:x="jelly:xml">
-    <st:documentation>
-        Generates the page header, along with its associated dynamic behaviours.
-        This tag is used by l:layout and not expected to be used by anyone else,
-        but it's written as separate tag for better readability of code.
-
-        <st:attribute name="title" required="true">
-            Page title and title attribute for the logo
-        </st:attribute>
-
-        <st:attribute name="logoAlt" required="true">
-            Alt text for the logo
-        </st:attribute>
-
-        <st:attribute name="searchPlaceholder" required="true">
-            Placeholder text for the search input
-        </st:attribute>
-
-        <st:attribute name="searchHelpUrl" required="true">
-            Link value for the help icon on the search box
-        </st:attribute>
-
-        <st:attribute name="logout" required="true">
-            Text for the logout link
-        </st:attribute>
-    </st:documentation>
-<<<<<<< HEAD
-	<j:invokeStatic var="header" className="jenkins.views.JenkinsHeader" method="get"/>
-	<st:include it="${header}" page="headerContent.jelly"/>
-=======
-
-    <header id="header" class="page-header">
-        <div class="page-header__brand">
-            <div class="logo">
-                <a id="jenkins-home-link" href="${rootURL}/">
-                    <img id="jenkins-head-icon" src="${imagesURL}/svgs/logo.svg" alt="[${logoAlt}]" />
-                    <img id="jenkins-name-icon" src="${imagesURL}/title.svg" alt="${title}" width="139" height="34" />
-                </a>
-            </div>
-
-            <a class="page-header__brand-link" href="${rootURL}/">
-                <img src="${imagesURL}/svgs/logo.svg"
-                     alt="[${logoAlt}]"
-                     class="page-header__brand-image" />
-                <span class="page-header__brand-name">Jenkins</span>
-            </a>
-        </div>
-
-        <div class="searchbox hidden-xs">
-            <!-- search box -->
-            <j:set var="searchURL" value="${h.searchURL}"/>
-            <form action="${searchURL}" method="get" style="position:relative;" class="no-json" name="search" role="search">
-                <!-- this div is used to calculate the width of the text box -->
-                <div id="search-box-sizer"/>
-                <div id="searchform">
-                    <input name="q" placeholder="${searchPlaceholder}" id="search-box" class="main-search__input" value="${request.getParameter('q')}" role="searchbox" />
-
-                    <span class="main-search__icon-leading">
-                        <l:svgIcon href="${resURL}/images/material-icons/svg-sprite-action-symbol.svg#ic_search_24px" />
-                    </span>
-                    <a href="${searchHelpUrl}" class="main-search__icon-trailing">
-                        <l:svgIcon href="${resURL}/images/material-icons/svg-sprite-action-symbol.svg#ic_help_outline_24px" />
-                    </a>
-
-                    <div id="search-box-completion" />
-                    <script>createSearchBox("${searchURL}");</script>
-                </div>
-            </form>
-        </div>
-
-        <div class="login page-header__hyperlinks">
-            <div id="visible-am-insertion" class="page-header__am-wrapper" />
-            <div id="visible-sec-am-insertion" class="page-header__am-wrapper" />
-
-            <!-- login field -->
-            <j:if test="${app.useSecurity}">
-                <j:choose>
-                    <j:when test="${!h.isAnonymous()}">
-                        <j:invokeStatic var="user" className="hudson.model.User" method="current" />
-                        <j:choose>
-                            <j:when test="${user.fullName == null || user.fullName.trim().isEmpty()}">
-                                <j:set var="userName" value="${user.id}"/>
-                            </j:when>
-                            <j:otherwise>
-                                <j:set var="userName" value="${user.fullName}"/>
-                            </j:otherwise>
-                        </j:choose>
-                        <a href="${rootURL}/${user.url}" class="model-link inside inverse">
-                            <l:svgIcon
-                                       class="am-monitor-icon" >
-                                <use href="${resURL}/images/material-icons/svg-sprite-social-symbol.svg#ic_person_24px"></use>
-                            </l:svgIcon>
-                            <span class="hidden-xs hidden-sm">${userName}</span>
-                        </a>
-                        <j:if test="${app.securityRealm.canLogOut()}">
-                            <a href="${rootURL}/logout">
-                                <l:svgIcon href="${resURL}/images/material-icons/svg-sprite-action-symbol.svg#ic_input_24px" />
-                                <span class="hidden-xs hidden-sm">${logout}</span>
-                            </a>
-                        </j:if>
-                    </j:when>
-                    <j:otherwise>
-                        <st:include it="${app.securityRealm}" page="loginLink.jelly" />
-                    </j:otherwise>
-                </j:choose>
-            </j:if>
-        </div>
-    </header>
->>>>>>> upstream/master
-</j:jelly>
